Move WorksIndex to ingestor
