+++ catalogue_pipeline/ingestor/src/main/scala/uk/ac/wellcome/platform/ingestor/models/WorksIndex.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.elasticsearch.mappings
+package uk.ac.wellcome.platform.ingestor.models
 
 import com.google.inject.Inject
 import com.sksamuel.elastic4s.analyzers._
