+++ catalogue_pipeline/ingestor/src/test/scala/uk/ac/wellcome/platform/ingestor/IngestorIndexTest.scala
@@ -4,10 +4,11 @@ import com.sksamuel.elastic4s.ElasticDsl.{deleteIndex, index}
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{FunSpec, Matchers}
 import com.sksamuel.elastic4s.http.ElasticDsl._
+import uk.ac.wellcome.platform.ingestor.test.utils.Ingestor
 
 class IngestorIndexTest
     extends FunSpec
-    with IngestorUtils
+    with Ingestor
     with Matchers
     with ScalaFutures {
