+++ catalogue_pipeline/ingestor/src/test/scala/uk/ac/wellcome/platform/ingestor/models/WorksIndexTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.elasticsearch.mappings
+package uk.ac.wellcome.platform.ingestor.models
 
 import com.sksamuel.elastic4s.http.ElasticDsl._
 import com.sksamuel.elastic4s.mappings.dynamictemplate.DynamicMapping
@@ -6,7 +6,7 @@ import org.elasticsearch.client.ResponseException
 import org.scalatest.concurrent.{Eventually, ScalaFutures}
 import org.scalatest.{BeforeAndAfterEach, FunSpec, Matchers}
 import uk.ac.wellcome.models.{IdentifierSchemes, _}
-import uk.ac.wellcome.test.utils.ElasticSearchLocal
+import uk.ac.wellcome.platform.ingestor.test.utils.ElasticSearchLocal
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
 import uk.ac.wellcome.utils.JsonUtil
